configs: mt8183: Rename MT8183 kernel config to be arch-generic
The kernel config for MT8183 can already boot on Hana boards. The lowest
of depthcharge size limits on ARM64 is a comfortable 32MiB, so we can
try to create a arch-generic kernel, and probably even an arch-generic
image. Rename the config and its references to "a64" as in the Makefile.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>